Extract `isDisabled()` into separate file
