+++ lib/assertions.js
@@ -7,6 +7,7 @@ import isRequired from './assertions/is-required';
 import isNotRequired from './assertions/is-not-required';
 import isVisible from './assertions/is-visible';
 import isNotVisible from './assertions/is-not-visible';
+import isDisabled from './assertions/is-disabled';
 import isNotDisabled from './assertions/is-not-disabled';
 
 import elementToString from './helpers/element-to-string';
@@ -318,24 +319,7 @@ export default class DOMAssertions {
    * @see {@link #isNotDisabled}
    */
   isDisabled(message) {
-    let element = this.findTargetElement();
-
-    let expected = `Element ${this.targetDescription} is disabled`;
-    let actual = `Element ${this.targetDescription} is disabled`;
-    let result = element.disabled;
-
-    if (result === false) {
-      actual = `Element ${this.targetDescription} is not disabled`;
-    } else if(result === undefined) {
-      actual = `Element ${this.targetDescription} does not support disabled`;
-      result = false;
-    }
-
-    if (!message) {
-      message = expected;
-    }
-
-    this.pushResult({ result, actual, expected, message });
+    isDisabled.call(this, message);
   }
 
   /**
